add jwt toeken

+++ src/posts/post-service.ts
@@ -1,8 +1,8 @@
-import { Injectable, NotFoundException } from "@nestjs/common";
+import { Injectable, NotFoundException, UnauthorizedException } from "@nestjs/common";
 import { InjectModel } from "@nestjs/sequelize";
 import { Post } from "./post-model";
 import { User } from "../User/user-model";
-import { UserRepository } from "src/user/user-repository";
+import { UserRepository } from "src/User/user-repository";
 import { PostRepository } from "./post-repository";
 
 
@@ -10,13 +10,13 @@ import { PostRepository } from "./post-repository";
 export class PostService {
     constructor(
         private postRepository: PostRepository,
-        private userRepository: UserRepository, // Inject UserRepository
+        private userRepository: UserRepository, 
     ) {}
 
 
 
     async createpost(postData:Partial<Post>):Promise<Post>{
-       const user=await this.userRepository.singleUser(postData.user_uuid);
+       const user=await this.userRepository.findById(postData.user_uuid);
         if (!user){
             throw new NotFoundException('User not found');
             } 
@@ -35,18 +35,32 @@ export class PostService {
 
     }
 
-    async updatePost(id:string,PostData:Partial<Post>):Promise<Post>{
+    async updatePost(id: string, postData: Partial<Post>, userId: string): Promise<Post> {
         const post = await this.postRepository.findById(id);
         if (!post) {
             throw new NotFoundException('Post not found');
         }
-        return this.postRepository.update_post(id, PostData);
+       
+        if (post.user_uuid !== userId) {
+            throw new UnauthorizedException('You are not authorized to edit this post');
+        }
+        return this.postRepository.update_post(id, postData);
     }
-    async deletPost(id:string):Promise<void>{
+    
+    
+    async deletePost(id: string, userId: string): Promise<void> {
+        const post = await this.postRepository.findById(id);
+        if (!post) {
+            throw new NotFoundException('Post not found');
+        }
+        
+        if (post.user_uuid !== userId) {
+            throw new UnauthorizedException('You are not authorized to delete this post');
+        }
         await this.postRepository.delete(id);
+    
+    
+
+
     }
-
-
-
-
 }
